Remove duplicate methods, fix more phpdocs.
