+++ src/Url.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace Yii\Extension\Form;
 
 use InvalidArgumentException;
-use Yii\Extension\Form\Attribute\InputAttributes;
-use Yii\Extension\Form\Contract\HasLengthContract;
+use Yii\Extension\Form\Attribute\RulesAttributes;
 use Yii\Extension\Form\Contract\PlaceholderContract;
-use Yii\Extension\Form\Contract\RegexContract;
 use Yiisoft\Html\Tag\Input;
 
 use function is_string;
@@ -19,29 +17,8 @@ use function is_string;
  *
  * @link https://www.w3.org/TR/2012/WD-html-markup-20120329/input.url.html
  */
-final class Url extends InputAttributes implements HasLengthContract, PlaceholderContract, RegexContract
+final class Url extends RulesAttributes implements PlaceholderContract
 {
-    public function maxlength(int $value): self
-    {
-        $new = clone $this;
-        $new->attributes['maxlength'] = $value;
-        return $new;
-    }
-
-    public function minlength(int $value): self
-    {
-        $new = clone $this;
-        $new->attributes['minlength'] = $value;
-        return $new;
-    }
-
-    public function pattern(string $value): self
-    {
-        $new = clone $this;
-        $new->attributes['pattern'] = $value;
-        return $new;
-    }
-
     public function placeholder(string $value): self
     {
         $new = clone $this;
